Change type2

+++ src/main/java/functions/Calculator2.java
@@ -3,7 +3,7 @@ package functions;
 import static java.lang.StrictMath.pow;
 
 public class Calculator2 extends Calculator {
-    public void power(int x) {
+    public void power(double x) {
         if (answer==null) return;//ignore
         answer = pow(answer, x);
     }
